chore(build): drop -fpermissive / -Wno-narrowing / -Wno-return-type / -Wno-error (soh, libultraship, ZAPDTR, OTRExporter); fix underlying narrowing sites
Per standing project policy (no -fpermissive / -Wno-<correctness>): these flags
mask real conformance bugs and are a documented source of parity divergence
(silent narrowing). Removed from soh, libultraship, ZAPDTR, and OTRExporter, and
fixed the two real narrowing sites they were hiding:

- Shipwright/ZAPDTR/ZAPD/ZRom.cpp:77 — IsMQ() read the ROM CRC into `int`, but
  OOT/MM CRC constants (e.g. 0xEC7011B7) exceed 0x7FFFFFFF and don't match a
  signed `case` label without narrowing. Store as `uint32_t` at the read site.

- Shipwright/soh/soh/Enhancements/ExtraModes/EnemyRandomizer.cpp:75-78 — the
  "Invisible <enemy>" entries store 0x8000/0x8001/0x8002/0x8004 as the actor
  params field (declared `s16`); the top-bit sentinel is intentional, so wrap
  each in an explicit static_cast<s16> to name the intent and satisfy -Wnarrowing.

- Shipwright/soh/soh/SaveManager.cpp:882/889/896/1005/1012/1019 — six
  `std::array<u8, 8> sPlayerName` initializers used `'\xNN'` char literals; on
  x86 `char` is signed, so `'\xB6'` etc. narrow when initializing a u8 element.
  Rewrite the six initializers with numeric hex literals (0xNN) — same bytes,
  no narrowing.

STOPGAP: -fpermissive and -Wno-narrowing kept ONLY in Shipwright/mm/CMakeLists.txt,
each with an inline comment naming the reason. MM's 2s2h BenGui port has
"declaration changes meaning" C++ conformance errors across UIWidgets*Options
(method names collide with in-scope enum types) plus many narrowing sites in
vendored code — fixing them is a rename refactor across BenGui and a separate
audit, not this change. Follow-up card should track dropping these two flags
from mm too.

Verified: ninja soh mm libultraship builds clean (0 errors). charcompare
pre-existing bitrot per memory [charcompare deprecated] — not gated on.

+++ Shipwright/soh/soh/Enhancements/ExtraModes/EnemyRandomizer.cpp
@@ -72,10 +72,10 @@ static EnemyEntry randomizedEnemySpawnTable[] = {
     { CVAR_ENHANCEMENT("RandomizedEnemyList.GohmaLarva"),       "Gohma Larva",           ACTOR_EN_GOMA,             7 }, // Gohma Larva (Non-Gohma rooms)
     { CVAR_ENHANCEMENT("RandomizedEnemyList.Guay"),             "Guay",                  ACTOR_EN_CROW,             0 }, // Guay
     { CVAR_ENHANCEMENT("RandomizedEnemyList.IceKeese"),         "Ice Keese",             ACTOR_EN_FIREFLY,          4 }, // Ice Keese
-    { CVAR_ENHANCEMENT("RandomizedEnemyList.InvisFireKeese"),   "Invisible Fire Keese",  ACTOR_EN_FIREFLY,     0x8001 }, // Fire Keese (invisible)
-    { CVAR_ENHANCEMENT("RandomizedEnemyList.InvisFloormaster"), "Invisible Floormaster", ACTOR_EN_FLOORMAS,    0x8000 }, // Floormaster (invisible)
-    { CVAR_ENHANCEMENT("RandomizedEnemyList.InvisIceKeese"),    "Invisible Ice Keese",   ACTOR_EN_FIREFLY,     0x8004 }, // Ice Keese (invisible)
-    { CVAR_ENHANCEMENT("RandomizedEnemyList.InvisKeese"),       "Invisible Keese",       ACTOR_EN_FIREFLY,     0x8002 }, // Keese (invisible)
+    { CVAR_ENHANCEMENT("RandomizedEnemyList.InvisFireKeese"),   "Invisible Fire Keese",  ACTOR_EN_FIREFLY,     static_cast<s16>(0x8001) }, // Fire Keese (invisible)
+    { CVAR_ENHANCEMENT("RandomizedEnemyList.InvisFloormaster"), "Invisible Floormaster", ACTOR_EN_FLOORMAS,    static_cast<s16>(0x8000) }, // Floormaster (invisible)
+    { CVAR_ENHANCEMENT("RandomizedEnemyList.InvisIceKeese"),    "Invisible Ice Keese",   ACTOR_EN_FIREFLY,     static_cast<s16>(0x8004) }, // Ice Keese (invisible)
+    { CVAR_ENHANCEMENT("RandomizedEnemyList.InvisKeese"),       "Invisible Keese",       ACTOR_EN_FIREFLY,     static_cast<s16>(0x8002) }, // Keese (invisible)
     { CVAR_ENHANCEMENT("RandomizedEnemyList.InvisRedead"),      "Invisible Redead",      ACTOR_EN_RD,               3 }, // Redead (invisible)
     { CVAR_ENHANCEMENT("RandomizedEnemyList.InvisSkulltula"),   "Invisible Skulltula",   ACTOR_EN_ST,               2 }, // Skulltula (invisible)
     { CVAR_ENHANCEMENT("RandomizedEnemyList.InvisStalfos"),     "Invisible Stalfos",     ACTOR_EN_TEST,             0 }, // Stalfos (invisible)
